Moves LogHandler and LogLevel to the :core module (#193)
